+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 
 * Add option to not install plugins
 * Add Python 3.5 to build matrix
+* Add argument to pass options to pip
 
 0.8.0 (2015-08-30)
 ++++++++++++++++++